requirements.txt
numpy==1.18.2
opencv-python==3.4.9.31
opencv-contrib-python==3.4.9.31
PyQt5==5.14.1
PyQt5-sip==12.7.1
PyYAML==5.3.1
PyQtWebEngine==5.14.1